refactor: 적금 계산 페이지 폴더로 이동
